backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
whoosh==2.7.4
numpy==1.26.4
faiss-cpu==1.8.0
google-generativeai==0.8.0
sse-starlette==2.1.0
